Change class layer toggle tool to be subclass of new general layer toggle tool.
