+++ src/hubbleds/tools/__init__.py
@@ -2,5 +2,5 @@ from .rest_wavelength_tool import RestWavelengthTool
 from .spectrum_flag_tool import SpectrumFlagTool
 from .wavelength_zoom import WavelengthZoom
 from .hubble_line_fit_tool import HubbleLineFitTool
-from .toggle_layer_tool import LayerToggleTool
+from .class_layer_toggle_tool import ClassLayerToggleTool
 from .hubble_line_draw_tool import HubbleLineDrawTool
